backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
slowapi==0.1.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Task queue
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Object storage
minio==7.2.7

# PDF generation
reportlab==4.2.2

# Scraping
playwright==1.44.0
scrapy==2.11.1
beautifulsoup4==4.12.3
pdfplumber==0.11.0
httpx==0.27.0
lxml==5.2.1

# ML
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
mlflow==2.12.1
shap==0.45.1

# Monitoring
prometheus-fastapi-instrumentator==7.0.0
sentry-sdk[fastapi]==2.3.1

# Scheduling
apscheduler==3.10.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0

# Dev tools
black==24.4.2
ruff==0.4.4
